wifi: update Wi-Fi applications for TizenRT 1.0 update
This patch includes updates on wpa_supplicant, wifi api and its header.
All the updates developed by S.LSI are fast-forwarded as well as p2p and
fsm although it has not been validated yet.

+++ apps/wpa_supplicant/src/l2_packet/l2_packet_tinyara.c
@@ -154,7 +154,7 @@ int l2_packet_get_ip_addr(struct l2_packet_data *l2, char *buf, size_t len)
 	}
 	os_memset(&ifr, 0, sizeof(ifr));
 	os_strlcpy(ifr.ifr_name, l2->ifname, sizeof(ifr.ifr_name));
-	if (ioctl(s, SIOCGIFADDR, &ifr) < 0) {
+	if (ioctl(s, SIOCGIFADDR, (unsigned long)&ifr) < 0) {
 		if (errno != EADDRNOTAVAIL) {
 			wpa_printf(MSG_ERROR, "%s: ioctl[SIOCGIFADDR]: %s", __func__, strerror(errno));
 		}
